Rename files using randwalk -> knots (#977b)
