+++ galsim/knots.py
